Refactor the expression handling to its own module
