+++ module/helpers/inline-helper.mjs
@@ -1,9 +1,8 @@
-import { FU, SYSTEM } from './config.mjs';
+import { SYSTEM } from './config.mjs';
 import { Flags } from './flags.mjs';
 import { FUActor } from '../documents/actors/actor.mjs';
 import { FUItem } from '../documents/items/item.mjs';
-import { ImprovisedEffect } from './improvised-effect.mjs';
-import { MathHelper } from './math-helper.mjs';
+import { Expressions } from '../expressions/expressions.mjs';
 
 /**
  * @description Information about a lookup for the source of an inline element
@@ -22,197 +21,6 @@ export class InlineSourceInfo {
 }
 
 /**
- * @property {FUActor} actor
- * @property {FUItem} item
- * @property {FUActor[]} targets
- */
-export class InlineContext {
-	constructor(actor, item, targets) {
-		this.actor = actor;
-		this.item = item;
-		this.targets = targets;
-	}
-}
-
-// DSL supported by the inline amount expression
-const referenceSymbol = '@';
-const actorLabel = `actor`;
-const itemLabel = `item`;
-
-/**
- * @property {String} text The raw text
- * @property {Boolean} dynamic Whether the amount needs to be evaluated based on the context
- */
-export class InlineAmount {
-	constructor(text) {
-		this.text = text;
-		this.dynamic = InlineAmount.isDynamic(text);
-	}
-
-	/**
-	 * @param text The raw text of the amount
-	 * @returns {boolean} True if the amount is dynamic
-	 */
-	static isDynamic(text) {
-		return !Number.isInteger(Number(text));
-	}
-
-	/**
-	 * @param {HTMLAnchorElement} anchor
-	 * @param {String} amount
-	 */
-	static appendToAnchor(anchor, amount) {
-		anchor.dataset.amount = amount;
-		const dynamicAmount = InlineAmount.isDynamic(amount);
-		if (dynamicAmount) {
-			anchor.append(game.i18n.localize('FU.Variable'));
-		} else {
-			anchor.append(amount);
-		}
-	}
-
-	/**
-	 * @param {InlineContext} context
-	 * @return {Number} The evaluated amount
-	 */
-	evaluate(context) {
-		if (!this.dynamic) {
-			return Number(this.text);
-		}
-
-		let expression = this.text;
-
-		// Evaluate the expression
-		let substitutedExpression = evaluateProperties(expression, context);
-		substitutedExpression = evaluateFunctions(substitutedExpression, context);
-		const result = MathHelper.evaluate(substitutedExpression);
-
-		if (Number.isNaN(result)) {
-			throw new Error(`Failed to evaluate expresson ${substitutedExpression}`);
-		}
-
-		console.info(`Substituted expression ${expression} > ${substitutedExpression} > ${result}`);
-		return result;
-	}
-}
-
-function evaluateFunctions(expression, context) {
-	const pattern = /\$(?<label>[a-zA-Z]+)\.(?<path>(\w+\.?)+)\((?<args>.*?)\)/gm;
-	function evaluate(match, label, path, p3, args, groups) {
-		if (match.includes(actorLabel)) {
-			if (context.actor == null) {
-				ui.notifications.warn('FU.ChatEvaluateAmountNoActor', { localize: true });
-				throw new Error(`No reference to an actor provided for "${match}"`);
-			}
-
-			let splitArgs = args.split(',');
-
-			const functionPath = `system.${path}`;
-			const resolvedFunction = getFunctionFromPath(context.actor, functionPath);
-			if (resolvedFunction === undefined) {
-				throw new Error(`No function in path "${functionPath}" of object ${context.actor}`);
-			}
-			const result = resolvedFunction.apply(context.actor.system, splitArgs);
-			console.info(`Resolved function ${functionPath}: ${result}`);
-			return result;
-		}
-	}
-	return expression.replace(pattern, evaluate);
-}
-
-function evaluateProperties(expression, context) {
-	const pattern = /(?<variable>@?([a-zA-Z]+\.?)+)/gm;
-	function evaluate(match, label, path, pN, offset, string, groups) {
-		// ImprovisedEffect
-		if (match in FU.improvisedEffect) {
-			return ImprovisedEffect.calculateAmountFromContext(match, context);
-		}
-		// Property Reference
-		else if (match.includes(referenceSymbol)) {
-			// TODO: Refactor
-			let root = null;
-			let propertyPath = '';
-
-			if (match.includes(itemLabel)) {
-				if (context.item == null) {
-					ui.notifications.warn('FU.ChatEvaluateAmountNoItem', { localize: true });
-					throw new Error(`No reference to an item provided for "${match}"`);
-				}
-				root = context.item;
-				propertyPath = match.replace(`${referenceSymbol}${itemLabel}.`, 'system.');
-			} else if (match.includes(actorLabel)) {
-				if (context.actor == null) {
-					ui.notifications.warn('FU.ChatEvaluateAmountNoActor', { localize: true });
-					throw new Error(`No reference to an actor provided for "${match}"`);
-				}
-				root = context.actor;
-				propertyPath = match.replace(`${referenceSymbol}${actorLabel}.`, 'system.');
-			}
-
-			// Evaluate the property value
-			const propertyValue = getPropertyValueByPath(root, propertyPath);
-			if (propertyValue === undefined) {
-				throw new Error(`Unexpected variable "${propertyPath}" in object ${root}`);
-			}
-			return propertyValue;
-		}
-		return match;
-	}
-
-	return expression.replace(pattern, evaluate);
-}
-
-/**
- * @param obj The object to resolve the function  from
- * @param path The path to the function, in dot notation
- * @returns {Function} The resolved function
- */
-function getFunctionFromPath(obj, path) {
-	if (typeof path !== 'string') {
-		throw new Error('Path must be a string');
-	}
-	if (typeof obj !== 'object' || obj === null) {
-		throw new Error('Invalid object provided');
-	}
-
-	const parts = path.split('.');
-	let current = obj;
-
-	for (const part of parts) {
-		if (current[part] === undefined) {
-			throw new Error(`Path not found in ${obj}: ${path}`);
-		}
-		current = current[part];
-	}
-
-	if (typeof current !== 'function') {
-		throw new Error(`Path does not resolve to a function: ${path}`);
-	}
-
-	return current;
-}
-
-/**
- * @param obj The object to resolve the property from
- * @param path The path to the property, in dot notation
- * @returns {undefined|*} The value of the property
- */
-function getPropertyValueByPath(obj, path) {
-	const keys = path.split('.');
-	let value = obj;
-
-	for (let key of keys) {
-		if (typeof value === 'object' && value !== null) {
-			value = value[key];
-		} else {
-			return undefined;
-		}
-	}
-
-	return value;
-}
-
-/**
  * @param {Document} document
  * @param {HTMLElement} element
  * @returns {InlineSourceInfo}
@@ -223,11 +31,9 @@ function determineSource(document, element) {
 	let actor = undefined;
 	let item = null;
 
-	// TODO: Make sure item always gets resolved
-	const itemId = $(element).closest('[data-item-id]').data('itemId');
-
 	if (document instanceof FUActor) {
 		actor = document;
+		const itemId = $(element).closest('[data-item-id]').data('itemId');
 		if (itemId) {
 			item = document.items.get(itemId);
 			name = item.name;
@@ -254,11 +60,25 @@ function determineSource(document, element) {
 		}
 	}
 
+	// TODO: Make sure item always gets resolved
 	return new InlineSourceInfo(name, uuid, actor, item);
 }
 
+/**
+ * @param {HTMLAnchorElement} anchor
+ * @param {String} amount
+ */
+function appendAmountToAnchor(anchor, amount) {
+	anchor.dataset.amount = amount;
+	const dynamicAmount = Expressions.requiresContext(amount);
+	if (dynamicAmount) {
+		anchor.append(game.i18n.localize('FU.Variable'));
+	} else {
+		anchor.append(amount);
+	}
+}
+
 export const InlineHelper = {
-	getPropertyValueByPath,
-	getFunctionFromPath,
 	determineSource,
+	appendAmountToAnchor,
 };
